Set `report_dashboard.last_viewed_at` on query
We don't really have a distinction between a card being "read" in the
sense that it's retrieved from the database, vs a card being "read" in
the sense that we're running its query. (The same is true for a
dashboard "read".)

So rather than reusing the `:event/card-read` or `:event/dashboard-read`
event, I'm publishing a separate event when a dashcard is executed. We
update the `report_dashboard.last_viewed_at` when handling this event.