Usage API
Playing around with a consolidated API for tracking usage of dashboards
and cards.

The idea here is that we offer a performant API for tracking card or
dashboard usage. `(usage/track! :model/Card card-id)` is all you need.

The module has the following public functions:

- `(column model)`: the column name for the last used at field. This is
what you want to use when you're trying to get the date of last usage
in a t2/select.

- `(track! model id)`: mentioned above, a performant API for tracking
usage of a card or dashboard. Internally, it adds to an in-memory queue
of jobs, deduplicating based on the model/id (selecting the latest usage
time available)

- `(init!)` and `(stop!)` for starting and stopping processing. By
default, processing happens in separate threads.

Playing around with performance, a single call to `track!` takes about
.005ms on my machine, so we could add 200,000 calls to `track!` in a
single HTTP request and only add one second to the response time.

On my machine, that example queue of 200,000 jobs would take ~1-2
seconds to fully process.

`init!` starts a worker thread that will check the queue every second.

`stop!` stops the workers and then drains the queue.