Allow getting non-model ancestors in search API
Previously we had a `model_ancestors` flag that allowed you to retrieve
the ancestors of the collection a "thing" was in, when you discovered it
in search.

This expands this a little, allowing you to get the collection ancestors
for any type of model you retrieve (assuming that makes sense).