pvt `collection-ids->honeysql-filter-clause`
Previously `collection` had two public functions:

`collection/permissions-set->visible-collection-ids` took a permissions
set and returned a set of collection IDs you should be allowed to see.

`collection/visible-collection-ids->honeysql-filter-clause` took the
output of the above and returned a honeysql filter clause that would
restrict `collection_id` (or another field you passed in) to be in the
set.

But the caller shouldn't need to know how exactly we're constructing the
honeysql filter clause. They just want the filter clause - deciding how
exactly to build it (get the set of collection IDs I can view, then
filter on it) is an implementation detail that we might move away from.
For example, hopefully we'll move to a permissions model where we can
essentially let the database figure out the set of IDs you have access
to, and the filter clause can use a subselect.