Some small perf improvements for coll perms
This is a minor performance improvement but probably worthwhile.

There are a few changes:

- add a new knob to `permissions-set->visible-collection-ids`. If an
`effective-child-of` is passed as part of the config map, it will only
return collection IDs that are an effective child of the passed parent.
Previously, we would find the complete set of visible collection IDs,
and then exclude things that weren't an effective child, which basically
involved passing *every* visible collection ID to the query *twice*:
once as `collection_id IN [visible-collection-ids]` and once as
`location NOT LIKE [a collection ID]` for a large subset of the visible
collection IDs. Now, when looking for effective children, we just find
the set of collection IDs that are effective children of the parent, and
only need to pass it off to SQL once.

- when turning a set of visible collection IDs to a honeysql filter
clause, check to see which is bigger: the set of visible collection IDs,
or the set of *invisible* collection IDs. Use the smaller set to create
the filter clause. This way, if you have access to 99.9% of collections,
we'll just pass the *remaining* .1% of collections to the query, like
`collection_id NOT IN [invisible-collection-ids]` instead of
`collection_id IN [visible-collection_ids]`.

Testing this out locally with 5000 collections, endpoints that
previously timed out with an error (too many SQL params) took ~1.5s to
return results. Still not ideal, but better.